ubi9 Compilation Warnings (#196)
* Initial openssl changes: replaced RSA with PKEY

* More changes related to deprecated encryption API change

* some code cleanup

* minor change

* remove unwanted changes

* ubi8:openssl-1.x and ubi9:openssl-3.x support

* include of param_build in condition

* minor changes for ubi8

* reverting ubi8 code to free RSA memory

* Initial changes to support encrypt/decrypt functionality for openSSL3.0

* Addressed review comment for failure of EVP functions

* Addressed review comments

* Couple of more warnings addressed for ubi9

* addressed warning from server_store module

* ubi8 compilation support

* ubi8 compilation support

+++ server_admin/src/janssonProcessPost.c
@@ -1725,7 +1725,7 @@ XAPI void ism_config_setServerName(int getLock, int setDefault) {
     struct utsname buf;
 
     /* get current admin port */
-	int port = ism_config_json_getAdminPort(getLock);
+	unsigned short port = ism_config_json_getAdminPort(getLock);
 
     /* Get current ServerName */
     if ( getLock == 1 ) pthread_rwlock_wrlock(&srvConfiglock);
